Удаление списка. Добавление логики для удаления списка и очищения списка на экране детальной информации списка. Рефактор кнопок new_item и save_item в list_menu.xml, изменение значения с ifRoom на always для того чтобы они пропали из всплывающего меню.

+++ app/src/main/java/com/ds_create/listofnotes/viewModels/MainViewModel.kt
@@ -47,7 +47,8 @@ class MainViewModel(database: MainDatabase): ViewModel() {
         dao.deleteNote(id)
     }
 
-    fun deleteListName(id: Int) = viewModelScope.launch {
-        dao.deleteListName(id)
+    fun deleteNoteList(id: Int, deleteList: Boolean) = viewModelScope.launch {
+       if (deleteList) dao.deleteListName(id)
+        dao.deleteListItemsById(id)
     }
 }
